refactored compiler utils into new file
